+++ crates/core/src/c/extract.h
@@ -3,8 +3,8 @@
 
 #include "histogram.h"
 
-extern size_t extract_palette(
-	Image img, size_t k, HistogramScratch hist, Histogram reps, u8 *out
+extern u32 extract_palette(
+	Image img, u32 k, HistogramScratch hist, Histogram reps, u8 *out
 );
 
 #endif
