+++ crates/core/src/c/histogram.h
@@ -4,14 +4,14 @@
 #include "types.h"
 
 typedef struct {
-	size_t threshold;
-	size_t *bins0, *bins1;
+	u32 threshold;
+	u64 *bins0, *bins1;
 	u8 *work, *aux;
 } HistogramScratch;
 
 typedef struct {
 	float *l, *a, *b, *w;
-	size_t len;
+	u32 len;
 } Histogram;
 
 extern void build_hist(Image img, HistogramScratch scratch, Histogram *out);
